[score] Add clang-tidy and clang-format aspects (#1)
* [score] Add clang-tidy and clang-format aspects

This PR adds clang-tidy and clang-format support via Bazel aspects.
In a nutshell, this allows users to run clang-tidy and clang-format
checks on their C/C++ code via Bazel targets.

For details, see the top level README.md and
the tool specific READMEs under
* quality/private/clang_tidy/README.md
* quality/private/clang_format/README.md

In the aspect part, the compiler configurations are extracted from
the build actions and passed to clang-tidy/clang-format tools.

The invocation of clang-tidy/clang-format is done via Python scripts
that are also part of this PR.

Original contributors in order of number of configurations:
* Sascha Moecker
* Stephan Muench
* Lucas Munaretto
* Jan Nowotsch
* Vitor Sorpile Geraldo
* Yuri Barboza
* Luiz Brito
* Nevin Xue
* Daniel Merget

Co-authored-by: Sascha Moecker <Sascha.Moecker@bmw.de>
Co-authored-by: Stephan Muench <Stephan.SM.Muench@bmw.de>
Co-authored-by: Lucas Munaretto <Lucas.LM.Munaretto@bmw.de>
Co-authored-by: Jan Nowotsch <Jan.Nowotsch@bmw.de>
Co-authored-by: Vitor Sorpile Geraldo <Vitor.Sorpile@partner.bmw.com.br>
Co-authored-by: Yuri Barboza <Yuri.Barboza@partner.bmw.com.br>
Co-authored-by: Luiz Brito <Luiz.BB.Brito@partner.bmw.com.br>
Co-authored-by: Nevin Xue <Nevin.Xue@batw.bmwgroup.com>
Co-authored-by: Daniel Merget <Daniel.Merget@bmw.de>

* [score] Temporarily remove workflows for docs + format

For the transition phase to move to the "does it all"
script `./scripts/run_all_tests.sh`.

---------

Co-authored-by: Stephan Muench <Stephan.SM.Muench@bmw.de>
Co-authored-by: Lucas Munaretto <Lucas.LM.Munaretto@bmw.de>
Co-authored-by: Jan Nowotsch <Jan.Nowotsch@bmw.de>
Co-authored-by: Vitor Sorpile Geraldo <Vitor.Sorpile@partner.bmw.com.br>
Co-authored-by: Yuri Barboza <Yuri.Barboza@partner.bmw.com.br>
Co-authored-by: Luiz Brito <Luiz.BB.Brito@partner.bmw.com.br>
Co-authored-by: Nevin Xue <Nevin.Xue@batw.bmwgroup.com>
Co-authored-by: Daniel Merget <Daniel.Merget@bmw.de>